% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignMatchingTargets.R
\name{alignMatchingTargets}
\alias{alignMatchingTargets}
\title{alignMatchingTargets}
\usage{
alignMatchingTargets(
  matched.targets.file = NULL,
  subset.start = "all",
  subset.end = "all",
  output.dir = NULL,
  algorithm = "localpair",
  adjust.direction = TRUE,
  remove.reverse.tag = TRUE,
  threads = 1,
  overwrite = FALSE,
  resume = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{matched.targets.file}{path to a folder of sequence alignments in phylip format.}

\item{subset.start}{contigs are added into existing alignment if algorithm is "add"}

\item{subset.end}{contigs are added into existing alignment if algorithm is "add}

\item{output.dir}{contigs are added into existing alignment if algorithm is "add"}

\item{algorithm}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{adjust.direction}{TRUE applies the adjust sequence direction function of MAFFT}

\item{remove.reverse.tag}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{threads}{path to a folder of sequence alignments in phylip format.}

\item{overwrite}{TRUE to overwrite a folder of samples with output.dir}

\item{resume}{TRUE to skip samples already completed}

\item{quiet}{TRUE to supress screen output}

\item{cleanup.files}{give a save name if you wnat to save the summary to file.}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for gather summary statistics on your alignment. Can be used for filtering or summarizing data.
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
