% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.assessPassFail.R
\name{alignmentAssess}
\alias{alignmentAssess}
\title{alignmentAssess}
\usage{
alignmentAssess(
  alignment = NULL,
  max.alignment.gap.percent = 0,
  min.taxa.alignment = 0,
  min.alignment.length = 0
)
}
\arguments{
\item{alignment}{alignment in DNAStringSet Format}

\item{max.alignment.gap.percent}{maximum percentage of gaps (including N and ?) permitted
across the whole alignment for it to pass. Alignments at or above this threshold fail.}

\item{min.taxa.alignment}{minimum number of sequences required for the alignment to pass.
Alignments with this many or fewer sequences fail.}

\item{min.alignment.length}{minimum alignment length (width in bp) required to pass.
Alignments at or below this length fail.}
}
\value{
a logical value, TRUE or FALSE, if your alignment passes (TRUE) or fails (FALSE) the filters
}
\description{
Function for assessing whether an entire alignment meets certain criteria
}
\examples{
\dontrun{
pass <- alignmentAssess(alignment = my.alignment,
                        max.alignment.gap.percent = 50,
                        min.taxa.alignment = 4,
                        min.alignment.length = 100)
}

}
