% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.alignmentConversion.R
\name{alignmentConversion}
\alias{alignmentConversion}
\title{alignmentConversion}
\usage{
alignmentConversion(input.alignment = NULL, end.format = NULL)
}
\arguments{
\item{input.alignment}{alignment file read into R in almost any format}

\item{end.format}{the format to convert the alignment to. Options include: DNABin, DNAStringSet, matrix, list}
}
\value{
returns a new alignment from the desired format
}
\description{
Function for easily converting alignments between different R alignment classes
}
\examples{
\dontrun{
# Convert a DNAStringSet alignment to a DNAbin matrix
dnabin.align <- alignmentConversion(input.alignment = my.stringset, end.format = "DNAbin")
}

}
