% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.concatenateAlignments.R
\name{concatenateAlignments}
\alias{concatenateAlignments}
\title{concatenateAlignments}
\usage{
concatenateAlignments(
  alignment.folder = NULL,
  output.name = NULL,
  partition.file = TRUE,
  output.format = c("phylip", "nexus", "fasta", "all"),
  partition.format = c("raxml", "table", "all"),
  overwrite = FALSE
)
}
\arguments{
\item{alignment.folder}{folder that contains aligmnents to be concatenated}

\item{output.name}{output file name}

\item{partition.file}{TRUE to save a partition file}

\item{output.format}{output file format. "all" to save all three types phylip, nexus, and fasta}

\item{partition.format}{partition file format. "all" to save both raxml and table format}

\item{overwrite}{TRUE to overwrite file. Default FALSE.}
}
\value{
saves to file concatenated alignments and partition files delimiting the coordinates of each indidividual marker
}
\description{
Function for concatenating a large number of alignments
}
\examples{
\dontrun{
concatenateAlignments(alignment.folder = "alignments/trimmed",
                      output.name = "concatenated_dataset",
                      partition.file = TRUE,
                      output.format = "phylip",
                      partition.format = "raxml",
                      overwrite = TRUE)
}

}
