% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.countGaps.R
\name{countAlignmentGaps}
\alias{countAlignmentGaps}
\title{countAlignmentGaps}
\usage{
countAlignmentGaps(alignment = NULL)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}
}
\value{
returns a named vector of the number of gaps, total number of basepairs, and percent gaps in an entire alignment
}
\description{
Function for counting the total number of gaps across the entire alignment
}
\examples{
\dontrun{
gap.stats <- countAlignmentGaps(alignment = my.alignment)
# returns a named vector: gaps, alignment (total sites), percent_gaps
}

}
