% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeLengthTable.R
\name{edgeLengthTable}
\alias{edgeLengthTable}
\title{edgeLengthTable}
\usage{
edgeLengthTable(tree = NULL, tips = FALSE)
}
\arguments{
\item{tree}{phylogenetic tree from ape read.tree in phylo format}

\item{tips}{whether to include the tip edges or not}
}
\value{
a data.frame is returned with the columns: "edge" = edge number; "node1" = ancestral node; and "node2" = descendent node. Node that node1 can be repeated because many descendents could have 1 ancestor.
}
\description{
This function creates an edge length table showing the two connecting nodes to the edge
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
edge.table = edgeLengthTable(tree = your.tree, tips = TRUE)

}
