% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatAlignmentFolder.R
\name{formatAlignmentFolder}
\alias{formatAlignmentFolder}
\title{formatAlignmentFolder}
\usage{
formatAlignmentFolder(
  alignment.path = NULL,
  out.dir = NULL,
  overwrite = FALSE,
  from.format = "nexus",
  to.format = "phylip"
)
}
\arguments{
\item{alignment.path}{a folder of individual alignments in nexus or phylip format}

\item{out.dir}{the name of the new directory with reformatted alignments}

\item{overwrite}{if TRUE overwrites file if it exists; FALSE the dataset is skipped}

\item{from.format}{the format to convert the alignment from (phylip, nexus, fasta)}

\item{to.format}{the format to convert the alignment to (phylip, nexus, fasta)}
}
\value{
saves to file a folder of alignments for the desired format
}
\description{
Function for easily converting alignments between different R alignment classes
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
