% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterozygousSites.R
\name{heterozygousSites}
\alias{heterozygousSites}
\title{heterozygousSites}
\usage{
heterozygousSites(alignment = NULL, count = TRUE, all = TRUE)
}
\arguments{
\item{alignment}{alignment in ape DNABin or a matrix format}

\item{count}{Whethe to return the count of parsimoney informative sites (TRUE) or the proportion (FALSE)}

\item{ambiguities}{Whether to consider ambiguities (TRUE) or not (FALSE)}
}
\value{
plots the phylogenetic tree and selected data associated with an AstralPlane object. Can optionally be saved to file as a PDF by giving save.file a file name.
}
\description{
Calculates the number or proportion of parsimony informative sites in an alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
