% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.makeFlankAlignments.R
\name{makeFlankAlignments}
\alias{makeFlankAlignments}
\title{makeFlankAlignments}
\usage{
makeFlankAlignments(
  alignment.directory = NULL,
  alignment.format = "phylip",
  output.directory = NULL,
  reference.type = c("target", "alignment"),
  reference.path = NULL,
  target.direction = TRUE,
  concatenate.intron.flanks = TRUE,
  threads = 1,
  memory = 1,
  overwrite = FALSE,
  mafft.path = NULL
)
}
\arguments{
\item{alignment.directory}{path to a folder of input sequence alignments}

\item{alignment.format}{format of the input alignments; currently "phylip" or "fasta"}

\item{output.directory}{path to the directory where flank alignments will be saved}

\item{reference.type}{whether to use a "target" fasta file or an "alignment" directory as the reference for locating the exon coordinates}

\item{reference.path}{path to the reference target fasta file (when reference.type = "target") or a directory of reference alignments (when reference.type = "alignment")}

\item{target.direction}{if TRUE, output alignments are oriented to match the direction of the reference sequence}

\item{concatenate.intron.flanks}{if TRUE, the left and right flanking regions are concatenated into a single alignment file; if FALSE they are saved as separate files with "_1" and "_2" suffixes}

\item{threads}{number of parallel processing threads}

\item{memory}{total memory in GB to allocate across threads}

\item{overwrite}{if TRUE, overwrite existing output files; if FALSE, skip alignments already present in the output directory}

\item{mafft.path}{system path to the MAFFT executable directory; NULL to use the system PATH}
}
\value{
saves flank alignments to output.directory in phylip format; nothing is returned to R
}
\description{
Extracts intron flank regions from a directory of whole-marker alignments by aligning a reference sequence (from a target fasta or a reference alignment directory) to each alignment with MAFFT, identifying the coordinates of the reference, and cutting out the flanking intronic sequence on either side. Flanks can be saved as two separate files or concatenated into a single alignment. Runs in parallel using foreach/doParallel.
}
