% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePolytomy.R
\name{makePolytomy}
\alias{makePolytomy}
\title{makePolytomy}
\usage{
makePolytomy(tree = NULL, polytomy.limit = NULL)
}
\arguments{
\item{tree}{phylogenetic tree from ape read.tree in phylo format}

\item{polytomy.limit}{the threshold for collapsing a node into a polytomy}
}
\value{
phylogenetic tree in phylo format, with the nodes collapsed into polytomies.
}
\description{
This function collapses nodes under a certain support value into a polytomy
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
new.tree = makePolytomy(tree = your.tree, polytomy.limit = 10)

}
