% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameTreeTips.R
\name{renameTreeTips}
\alias{renameTreeTips}
\title{renameTreeTips}
\usage{
renameTreeTips(
  tree.directory = NULL,
  tree.extension = ".tre",
  rename.file = NULL,
  recursive = TRUE
)
}
\arguments{
\item{tree.directory}{the directory of trees to have their samples renamed}

\item{tree.extension}{the file extension of your tree files to be named}

\item{rename.file}{a tab delimited text file with two columns named Old_Name and New_Name for the current and new name of the target sample}

\item{recursive}{TRUE to recursively rename within sub-directories or FALSE just for the main directory}
}
\value{
overwrites trees in directory with renamed sample trees
}
\description{
Function for easily renaming the tips of all phylogenetic trees within a directory (or a single tree)
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
