% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.replaceAlignmentCharacter.R
\name{replaceAlignmentCharacter}
\alias{replaceAlignmentCharacter}
\title{replaceAlignmentCharacter}
\usage{
replaceAlignmentCharacter(
  alignment = NULL,
  char.find = "N",
  char.replace = "-"
)
}
\arguments{
\item{alignment}{a DNAStringSet containing the aligned sequences to modify}

\item{char.find}{the character to search for and replace in all sequences}

\item{char.replace}{the character to substitute in place of char.find}
}
\value{
a DNAStringSet with all occurrences of char.find replaced by char.replace
}
\description{
Replaces all occurrences of a specified character with another character in every sequence of a DNAStringSet alignment. A common use is converting N characters to gap characters ("-") before trimming.
}
