% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeAlignments.R
\name{summarizeAlignments}
\alias{summarizeAlignments}
\title{summarizeAlignments}
\usage{
summarizeAlignments(
  alignment.path = NULL,
  file.export = NULL,
  overwrite = FALSE,
  dataset.name = NULL,
  alignment.format = c("phylip", "nexus")
)
}
\arguments{
\item{alignment.path}{path to a folder of sequence alignments in phylip format.}

\item{file.export}{give a save name if you wnat to save the summary to file.}

\item{overwrite}{if TRUE overwrites file if it exists; FALSE the dataset is skipped}

\item{dataset.name}{A name for your dataset. i.e. exons, introns, UCEs}

\item{alignment.type}{select the format of the alignment. Phylip is avaialble for now, will be expanded in the future.}
}
\value{
returns a data.table with the raw summary statistics calculated for each alignment in the alignment.path. A csv file can optionally be saved by giving a file name to file.export
}
\description{
Function for gather summary statistics on your alignment. Can be used for filtering or summarizing data.
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
