% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.trimAlignmentTargets.R
\name{trimAlignmentTargets}
\alias{trimAlignmentTargets}
\title{trimAlignmentTargets}
\usage{
trimAlignmentTargets(
  alignment.directory = NULL,
  alignment.format = "phylip",
  output.directory = NULL,
  target.file = NULL,
  target.direction = TRUE,
  min.alignment.length = 100,
  min.taxa.alignment = 4,
  threads = 1,
  memory = 1,
  overwrite = FALSE,
  mafft.path = NULL
)
}
\arguments{
\item{alignment.directory}{path to a directory of input sequence alignments}

\item{alignment.format}{format of the input alignments; "phylip" or "fasta"}

\item{output.directory}{path to the directory where target-trimmed alignments will be saved}

\item{target.file}{path to a fasta file containing the target/reference sequences; each sequence name must match the corresponding alignment file name (without extension)}

\item{target.direction}{if TRUE, output alignments are oriented to match the direction of the target sequence}

\item{min.alignment.length}{minimum alignment length in bp required to save an alignment after trimming}

\item{min.taxa.alignment}{minimum number of sequences required to save an alignment after trimming}

\item{threads}{number of parallel processing threads}

\item{memory}{total memory in GB to allocate across threads}

\item{overwrite}{if TRUE, overwrite an existing output directory; if FALSE, skip alignments already present in the output directory}

\item{mafft.path}{system path to the MAFFT executable directory; NULL to use the system PATH}
}
\value{
saves target-trimmed alignment files to output.directory; nothing is returned to R
}
\description{
Trims each alignment in a directory to the region spanned by a corresponding target sequence from a reference fasta file. For each alignment, the matching target sequence is added to the alignment with MAFFT, the target's aligned coordinates are identified, large internal gaps in the target are corrected, and the alignment is subsetted to the target region before saving. Alignments that are too short or have too few taxa after trimming are discarded. Runs in parallel using foreach/doParallel.
}
