% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimHmmCleaner.R
\name{trimHmmCleaner}
\alias{trimHmmCleaner}
\title{trimHmmCleaner}
\usage{
trimHmmCleaner(
  alignment = NULL,
  specificity = TRUE,
  large = FALSE,
  hmmcleaner.path = "HmmCleaner.pl",
  quiet = FALSE,
  delete.temp = TRUE
)
}
\arguments{
\item{alignment}{alignment in DNAStringSet format}

\item{specificity}{TRUE to enable to more sensitive but slower "specificity" hmmCleaner option}

\item{large}{TRUE to enable the faster "large" dataset option for large datasets. Will not run "large" if alignment has less than 50 sequences}

\item{hmmcleaner.path}{Absolute path to hmmCleaner if R cannot find it in your path}

\item{quiet}{TRUE to supress HmmCleaner screen output}

\item{delete.temp}{TRUE to delete temporary files made by HmmCleaner}
}
\value{
returns a data.table with the raw summary statistics calculated for each alignment in the alignment.path. A csv file can optionally be saved by giving a file name to file.export
}
\description{
wrapper function for running HmmCleaner. Must be installed.
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
