% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentAssess.R
\name{alignmentAssess}
\alias{alignmentAssess}
\title{alignmentAssess}
\usage{
alignmentAssess(
  alignment = NULL,
  min.gap.percent = 0,
  min.taxa.count = 0,
  min.align.length = 0
)
}
\arguments{
\item{alignment}{alignment in DNAStringSet Format}

\item{min.gap.percent}{minimum percentage of gaps allowed for an alignment to pass}

\item{min.taxa.count}{minimum number of taxa allowed for an alignment to pass}

\item{min.align.length}{minimum alignment length (width) allowed for an alignment to pass}
}
\value{
a logical value, TRUE or FALSE, if your alignment passes (TRUE) or fails (FALSE) the filters
}
\description{
Function for assessing whether an entire alignment meets certain criteria
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
