% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenateGenes.R
\name{concatenateGenes}
\alias{concatenateGenes}
\title{concatenateGenes}
\usage{
concatenateGenes(
  alignment.folder = NULL,
  output.folder = NULL,
  feature.gene.names = NULL,
  input.format = c("phylip", "nexus", "fasta"),
  output.format = c("phylip", "nexus", "fasta"),
  remove.reverse = FALSE,
  overwrite = FALSE,
  threads = 1,
  memory = 1
)
}
\arguments{
\item{alignment.folder}{folder that contains aligmnents to be concatenated}

\item{output.folder}{output file name}

\item{input.format}{input file format. Save three types: phylip, nexus, and fasta}

\item{output.format}{output file format. Save three types: phylip, nexus, and fasta}

\item{remove.reverse}{TRUE to remove "\emph{R}" placed before reversed sequences in some alignments. Default FALSE.}

\item{overwrite}{TRUE to overwrite file. Default FALSE.}

\item{threads}{TRUE to overwrite file. Default FALSE.}

\item{memory}{TRUE to overwrite file. Default FALSE.}

\item{exon.gene.names}{output file name}
}
\value{
saves to file concatenated alignments and partition files delimiting the coordinates of each indidividual marker
}
\description{
Function for concatenating a large number of alignments
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
