% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countAlignmentGaps.R
\name{countAlignmentGaps}
\alias{countAlignmentGaps}
\title{countAlignmentGaps}
\usage{
countAlignmentGaps(alignment = NULL)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}
}
\value{
returns a named vector of the number of gaps, total number of basepairs, and percent gaps in an entire alignment
}
\description{
Function for counting the total number of gaps across the entire alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
