% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropboxDownload.R
\name{dropboxDownload}
\alias{dropboxDownload}
\title{dropboxDownload}
\usage{
dropboxDownload(
  sample.spreadsheet = NULL,
  dropbox.directory = NULL,
  dropbox.token = NULL,
  out.directory = NULL,
  overwrite = TRUE
)
}
\arguments{
\item{sample.spreadsheet}{spreadsheet path or table of file names to download}

\item{dropbox.directory}{the directory that contains these target files}

\item{dropbox.token}{your dropbox token path, created the first time you run the function but needs interactivity for log-in. Can be done once on your computer and the token can be moved around.}

\item{out.directory}{the directory to save the files you are downloading}

\item{overwrite}{whether to overwrite or not}
}
\value{
saves dropbox files to output directory
}
\description{
Function for downloading data off your dropbox account
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
