% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConsensus.R
\name{makeConsensus}
\alias{makeConsensus}
\title{makeConsensus}
\usage{
makeConsensus(
  alignment = NULL,
  method = c("majority", "threshold", "IUPAC", "profile"),
  threshold = 0.6,
  warn.non.IUPAC = FALSE,
  type = c("DNA", "RNA")
)
}
\arguments{
\item{alignment}{alignment in DNAStringSet format}

\item{method}{method to apply; majority = majority consensus, threshold = consensus from threshold percent of column; IUPAC and profile}

\item{threshold}{threshold to use for "threshold" method}

\item{warn.non.IUPAC}{warn if characters are not IUPAC}

\item{type}{DNA or RNA sequence data}
}
\value{
returns a data.table with the raw summary statistics calculated for each alignment in the alignment.path. A csv file can optionally be saved by giving a file name to file.export
}
\description{
Function for making a consensus sequence from alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
