% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseDistanceTarget.R
\name{pairwiseDistanceTarget}
\alias{pairwiseDistanceTarget}
\title{pairwiseDistanceTarget}
\usage{
pairwiseDistanceTarget(alignment = NULL, target = NULL)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}

\item{target}{a single sequence or reference aligned to alignment (or the name in the alignment) to pairwise compare each sample}
}
\value{
returns DNAStringSet of trimmed alignment
}
\description{
Function for trimming out divergent sample sequence segments in an alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
