% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloProcessR-package.R
\docType{package}
\name{PhyloProcessR-package}
\alias{PhyloProcessR}
\alias{PhyloProcessR-package}
\title{PhyloProcessR: Targeted Sequence-Capture Phylogenomics Workflows}
\description{
PhyloProcessR provides workflow scripts and R functions for processing
targeted sequence-capture data from raw reads to curated phylogenomic
datasets. See the project README and workflow configuration files for the
external tools required by each stage.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chutter/PhyloProcessR}
  \item Report bugs at \url{https://github.com/chutter/PhyloProcessR/issues}
}

}
\author{
\strong{Maintainer}: Carl Hutter \email{carl.hutter@gmail.com} (\href{https://orcid.org/0000-0001-6381-6339}{ORCID})

}
\keyword{internal}
