% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAdaptors.R
\name{removeAdaptors}
\alias{removeAdaptors}
\title{removeAdaptors}
\usage{
removeAdaptors(
  raw.reads = NULL,
  file.rename = NULL,
  output.dir = NULL,
  mode = c("sample", "directory"),
  fastp.path = "fastp",
  threads = 1,
  mem = 8,
  resume = TRUE,
  overwrite = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{raw.reads}{path to a folder of raw reads in fastq format.}

\item{file.rename}{a csv file with a "File" and "Sample" columns, where "File" is the file name and "Sample" is the desired renamed file}

\item{output.dir}{the new directory to save the adaptor trimmed sequences}

\item{mode}{"Sample" to run on a single sample or "Directory" to run on a directory of samples}

\item{fastp.path}{system path to fastp in case it can't be found}

\item{threads}{number of computation processing threads}

\item{mem}{amount of system memory to use}

\item{resume}{TRUE to skip samples already completed}

\item{overwrite}{TRUE to overwrite a folder of samples with output.dir}

\item{quiet}{TRUE to supress screen output}
}
\value{
a new directory of adaptor trimmed reads and a summary of the trimming in logs/
}
\description{
Function for removing adaptor sequences from raw Illumina sequence data using the program fastp
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
