% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeContamination.R
\name{removeContamination}
\alias{removeContamination}
\title{removeContamination}
\usage{
removeContamination(
  input.reads = "adaptor-removed-reads",
  output.dir = "decontaminated-reads",
  decontamination.path = NULL,
  mode = c("sample", "directory"),
  map.match = 1,
  read.mapper = "bwa",
  bbmap.path = "bbsplit.sh",
  samtools.path = "samtools",
  bwa.path = "bwa",
  threads = 1,
  mem = 1,
  resume = TRUE,
  overwrite = FALSE,
  quiet = TRUE
)
}
\arguments{
\item{input.reads}{path to a folder of adaptor trimmed reads in fastq format.}

\item{output.dir}{the new directory to save the adaptor trimmed sequences}

\item{decontamination.path}{directory of genomes contaminants to scan samples}

\item{mode}{"Sample" to run on a single sample or "Directory" to run on a directory of samples}

\item{bbmap.path}{system path to bbmap in case it can't be found}

\item{samtools.path}{system path to samtools in case it can't be found}

\item{bwa.path}{system path to bwa in case it can't be found}

\item{threads}{number of computation processing threads}

\item{mem}{amount of system memory to use}

\item{resume}{TRUE to skip samples already completed}

\item{overwrite}{TRUE to overwrite a folder of samples with output.dir}

\item{quiet}{TRUE to supress screen output}
}
\value{
a new directory of adaptor trimmed reads and a summary of the trimming in logs/
}
\description{
Function for removing contamination from other organisms from adaptor trimmed Illumina sequence data using BWA
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
