% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMafft.R
\name{runMafft}
\alias{runMafft}
\title{runMafft}
\usage{
runMafft(
  sequence.data = NULL,
  add.contigs = NULL,
  algorithm = "localpair",
  adjust.direction = TRUE,
  save.name = NULL,
  threads = 1,
  cleanup.files = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{sequence.data}{path to a folder of sequence alignments in phylip format.}

\item{add.contigs}{contigs are added into existing alignment if algorithm is "add"}

\item{algorithm}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{adjust.direction}{TRUE applies the adjust sequence direction function of MAFFT}

\item{save.name}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{threads}{path to a folder of sequence alignments in phylip format.}

\item{cleanup.files}{give a save name if you wnat to save the summary to file.}

\item{quiet}{TRUE to supress mafft screen output}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for gather summary statistics on your alignment. Can be used for filtering or summarizing data.
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
