% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSpades.R
\name{runSpades}
\alias{runSpades}
\title{runSpades}
\usage{
runSpades(
  read.paths = NULL,
  full.path.spades = "spades.py",
  mismatch.corrector = FALSE,
  read.contigs = T,
  save.file = T,
  save.name = NULL,
  threads = 1,
  memory = 4,
  overwrite = T,
  quiet = T
)
}
\arguments{
\item{read.paths}{path to a folder of sequence alignments in phylip format.}

\item{full.path.spades}{contigs are added into existing alignment if algorithm is "add"}

\item{mismatch.corrector}{algorithm to use: "add" add sequences with "add.contigs"; "localpair" for local pair align. All others available}

\item{read.contigs}{TRUE applies the adjust sequence direction function of MAFFT}

\item{save.file}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{save.name}{if a file name is provided, save.name will be used to save aligment to file as a fasta}

\item{threads}{number of computation processing threads}

\item{overwrite}{TRUE to overwrite a folder of samples with output.dir}

\item{quiet}{TRUE to supress screen output}

\item{mem}{amount of system memory to use}

\item{resume}{TRUE to skip samples already completed}
}
\value{
an alignment of provided sequences in DNAStringSet format. Also can save alignment as a file with save.name
}
\description{
Function for running the program spades to assemble short read sequencing data
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
