% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimExternal.R
\name{trimExternal}
\alias{trimExternal}
\title{trimExternal}
\usage{
trimExternal(alignment = NULL, min.n.seq = 4, codon.trim = T)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}

\item{min.n.seq}{minimum number of sequences needed to keep external colums}

\item{codon.trim}{trim in triplets to avoid harming reading frame}
}
\value{
returns DNAStringSet of externally trimmed alignment
}
\description{
Function for externally trimming at the edges in an alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
