% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSampleCoverage.R
\name{trimSampleCoverage}
\alias{trimSampleCoverage}
\title{trimSampleCoverage}
\usage{
trimSampleCoverage(
  alignment = NULL,
  min.coverage.percent = 50,
  min.sample.bp = 60
)
}
\arguments{
\item{alignment}{folder that contains an alignment in DNAStringSet format}

\item{min.coverage.percent}{minimum percent sequence data present for each sample out of the total alignment length}

\item{min.sample.bp}{minimum number of base pairs a sample needs to be kept}
}
\value{
returns sample coverage trimmed alignment. Alignment will be returned unmodified if the min.sample.bp is less than the alignment length.
}
\description{
Function for trimming out low coverage (i.e. high gap) samples from alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
