% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSampleSegments.R
\name{trimSampleSegments}
\alias{trimSampleSegments}
\title{trimSampleSegments}
\usage{
trimSampleSegments(alignment = NULL, slice.size.bp = 100, threshold = 0.45)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}

\item{slice.size.bp}{size of each sliding window to calculate divergence}

\item{threshold}{pair-wise distance threshold from consensus to trim out segment}
}
\value{
returns DNAStringSet of trimmed alignment
}
\description{
Function for trimming out divergent sample sequence segments in an alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
