% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimSampleSimularity.R
\name{trimSampleSimilarity}
\alias{trimSampleSimilarity}
\title{trimSampleSimilarity}
\usage{
trimSampleSimilarity(
  alignment = NULL,
  similarity.threshold = 0.4,
  realign.mafft = TRUE
)
}
\arguments{
\item{alignment}{alignment in DNAbin, DNAStringSet, list, and matrix formats}

\item{similarity.threshold}{divergence threshold from consensus to remove a sample}

\item{realign.mafft}{TRUE to realign alignment with mafft after removing samples}
}
\value{
returns DNAStringSet of trimmed alignment
}
\description{
Function for trimming out divergent sample sequence segments in an alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)

}
