% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimTrimal.R
\name{trimTrimal}
\alias{trimTrimal}
\title{trimTrimal}
\usage{
trimTrimal(
  alignment = NULL,
  trimal.path = "trimal",
  method = "auto",
  quiet = TRUE
)
}
\arguments{
\item{alignment}{path to a folder of sequence alignments in phylip format.}

\item{trimal.path}{provide system absolute path to trimal if R cannot find it}

\item{method}{trimming method, "auto" is automated. Other options coming soon.}

\item{quiet}{TRUE to supress TrimAl screen output}
}
\value{
returns an alignment in DNAStringSet format
}
\description{
wrapper function for the program TrimAl that corrects for NAs introduced from TrimAl removing samples from the alignment
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
