% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFasta.R
\name{writeFasta}
\alias{writeFasta}
\title{writeFasta}
\usage{
writeFasta(
  sequences,
  names,
  file.out,
  open = "w",
  nbchar = 60,
  as.string = FALSE
)
}
\arguments{
\item{sequences}{the sequences in some format to do}

\item{names}{names from the sequences}

\item{file.out}{name of the file to be saved}

\item{open}{keep the connection open = a to append, or 'w" to write once}

\item{nbchar}{number of characters per line if interleaving}

\item{as.string}{save as a string or not}
}
\value{
saves the alignment as a fasta file
}
\description{
Function for writing sequences or alignment in R to fasta format
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
