% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePhylip.R
\name{writePhylip}
\alias{writePhylip}
\title{writePhylip}
\usage{
writePhylip(alignment = NULL, file = NULL, interleave = FALSE, strict = FALSE)
}
\arguments{
\item{alignment}{summary data file from filterSummary}

\item{file}{summary data file from alignmentSummary}

\item{interleave}{folder of alignments to be filtered}

\item{strict}{save format for alignments}
}
\value{
saves the alignment as a phylip file
}
\description{
Function for writing alignments in R to phylip format
}
\examples{

your.tree = ape::read.tree(file = "file-path-to-tree.tre")
astral.data = astralPlane(astral.tree = your.tree,
                          outgroups = c("species_one", "species_two"),
                          tip.length = 1)


}
